+++ android/src/main/kotlin/com/godwin/calendar_events/Result.kt
@@ -2,6 +2,6 @@ package com.godwin.calendar_events
 
 sealed interface CalenderListResult
 
-class CalenderListSuccess(val list: List<Map<String, Any>>) : CalenderListResult
+class CalenderListSuccess(val list: List<Map<String, Any?>>) : CalenderListResult
 class CalenderListFailed(val error: CalenderError) : CalenderListResult